Changing Class names so bundler auto-requires gem
