+++ spec/bambora/rest_client_spec.rb
@@ -3,21 +3,21 @@
 require 'spec_helper'
 
 module Bambora
-  describe Client do
+  describe RestClient do
     describe '#initialize' do
       context 'with defaults' do
-        subject { Bambora::Client.new }
+        subject { described_class.new }
 
         it 'instantiates' do
-          expect(subject).to be_a Bambora::Client
+          expect(subject).to be_a Bambora::RestClient
         end
       end
 
       context 'with a different version' do
-        subject { Bambora::Client.new(version: 'V2') }
+        subject { described_class.new(version: 'V2') }
 
         it 'returns a raises an error' do
-          expect { subject }.to(raise_error(Bambora::Error, 'Only V1 endpoints are supported at this time.'))
+          expect { subject }.to(raise_error(Bambora::Client::Error, 'Only V1 endpoints are supported at this time.'))
         end
       end
     end
